Move map_info test into test.rs
Let's co-locate the map_info test with the remaining end-to-end tests
for easier management.
