+++ libbpf-rs/src/map.rs
@@ -987,48 +987,3 @@ impl MapInfo {
         MapFlags::from_bits_truncate(self.info.map_flags as u64)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use crate::Map;
-    use crate::MapFlags;
-    use crate::MapInfo;
-    use crate::MapType;
-
-    /// Test whether `MapInfo` works properly
-    #[test]
-    pub fn test_map_info() {
-        let opts = libbpf_sys::bpf_map_create_opts {
-            sz: mem::size_of::<libbpf_sys::bpf_map_create_opts>() as libbpf_sys::size_t,
-            map_flags: libbpf_sys::BPF_ANY,
-            btf_fd: 0,
-            btf_key_type_id: 0,
-            btf_value_type_id: 0,
-            btf_vmlinux_value_type_id: 0,
-            inner_map_fd: 0,
-            map_extra: 0,
-            numa_node: 0,
-            map_ifindex: 0,
-        };
-
-        let map = Map::create(MapType::Hash, Some("simple_map"), 8, 64, 1024, &opts).unwrap();
-        let map_info = MapInfo::new(map.fd()).unwrap();
-        let name_received = map_info.name().unwrap();
-        assert_eq!(name_received, "simple_map");
-        assert_eq!(map_info.map_type(), MapType::Hash);
-        assert_eq!(map_info.flags() & MapFlags::ANY, MapFlags::ANY);
-
-        let map_info = &map_info.info;
-        assert_eq!(map_info.key_size, 8);
-        assert_eq!(map_info.value_size, 64);
-        assert_eq!(map_info.max_entries, 1024);
-        assert_eq!(map_info.btf_id, 0);
-        assert_eq!(map_info.btf_key_type_id, 0);
-        assert_eq!(map_info.btf_value_type_id, 0);
-        assert_eq!(map_info.btf_vmlinux_value_type_id, 0);
-        assert_eq!(map_info.map_extra, 0);
-        assert_eq!(map_info.ifindex, 0);
-    }
-}
